feat: trigger stacks by unhealthy status
