+++ cmd/terramate/e2etests/cloud_helpers_test.go
@@ -1,0 +1,41 @@
+// Copyright 2023 Terramate GmbH
+// SPDX-License-Identifier: MPL-2.0
+
+package e2etest
+
+import (
+	"errors"
+	"net/http"
+	"testing"
+	"time"
+
+	"github.com/terramate-io/terramate/cloud/testserver"
+)
+
+func startFakeTMCServer(t *testing.T) {
+	fakeserver := &http.Server{
+		Handler: testserver.Router(),
+		Addr:    "localhost:3001",
+	}
+
+	const fakeserverShutdownTimeout = 3 * time.Second
+	errChan := make(chan error)
+	go func() {
+		errChan <- fakeserver.ListenAndServe()
+	}()
+
+	t.Cleanup(func() {
+		err := fakeserver.Close()
+		if err != nil {
+			t.Logf("fakeserver HTTP Close error: %v", err)
+		}
+		select {
+		case err := <-errChan:
+			if err != nil && !errors.Is(err, http.ErrServerClosed) {
+				t.Error(err)
+			}
+		case <-time.After(fakeserverShutdownTimeout):
+			t.Error("time excedeed waiting for fakeserver shutdown")
+		}
+	})
+}
