+++ test/cloud/client.go
@@ -6,8 +6,10 @@ package cloud
 import (
 	"context"
 	"strconv"
+	"testing"
 	"time"
 
+	"github.com/madlambda/spells/assert"
 	"github.com/terramate-io/terramate/cloud"
 )
 
@@ -17,7 +19,7 @@ const defaultTestTimeout = 1 * time.Second
 
 // PutStack sets a new stack in the /v1/stacks/<org>/<stack id>.
 // Note: this is not a real endpoint.
-func PutStack(orgUUID string, st cloud.Stack) error {
+func PutStack(t *testing.T, orgUUID string, st cloud.Stack) {
 	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
 	defer cancel()
 
@@ -26,7 +28,7 @@ func PutStack(orgUUID string, st cloud.Stack) error {
 		Credential: &credential{},
 	}
 	_, err := cloud.Put[cloud.EmptyResponse](ctx, client, st, cloud.StacksPath, orgUUID, strconv.Itoa(st.ID))
-	return err
+	assert.NoError(t, err)
 }
 
 type credential struct{}
